+++ src/main/java/com/genesis/cryptowallet/dto/ComparePositionDTO.java
@@ -4,12 +4,14 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
 public class ComparePositionDTO {
 
     private String symbol;
-    private double currentPosition;
-    private double priceCompared;
+    private BigDecimal currentPosition;
+    private BigDecimal priceCompared;
 }
